requirements.txt
pandas
requests
python-dotenv
psycopg2-binary
openai
boto3
resend
fastapi
uvicorn
python-multipart
openpyxl
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyJWT
cryptography
fpdf2
requests
